+++ rest_live/consumers.py
@@ -2,10 +2,11 @@ from typing import Any, Dict, Type, List, Union, Set
 
 from asgiref.sync import async_to_sync
 from channels.generic.websocket import JsonWebsocketConsumer
-from django.db.models import Model
 
-from rest_live import DEFAULT_GROUP_BY_FIELD, get_group_name, DELETED
+from rest_live import get_group_name, DELETED
 from rest_live.mixins import RealtimeMixin
+
+KwargType = Dict[str, Union[int, str]]
 
 
 class SubscriptionConsumer(JsonWebsocketConsumer):
@@ -15,8 +16,8 @@ class SubscriptionConsumer(JsonWebsocketConsumer):
         self.scope["method"] = "GET"
         self.subscriptions: Dict[str, List[int]] = dict()  # Maps group name to list of request IDs
         self.actions: Dict[int, str] = dict()  # Request ID to action (list or retrieve)
-        self.kwargs: Dict[int, Dict[str, Union[int, str]]] = dict()  # Request ID to view kwargs
-        self.pks: Dict[int, Set[int]] = dict()
+        self.kwargs: Dict[int, KwargType] = dict()  # Request ID to view kwargs
+        self.owned_instance_pks: Dict[int, Set[int]] = dict()
         self.accept()
 
     def add_subscription(self, group_name, request_id, **kwargs):
@@ -120,7 +121,7 @@ class SubscriptionConsumer(JsonWebsocketConsumer):
             group_name = get_group_name(model_label)
             self.add_subscription(group_name, request_id, **kwargs)
             self.actions[request_id] = viewset_action
-            self.pks[request_id] = viewset.get_list_pks()
+            self.owned_instance_pks[request_id] = viewset.get_list_pks()
         elif message_type == "unsubscribe":
             self.remove_subscription(request_id)
         else:
@@ -129,7 +130,6 @@ class SubscriptionConsumer(JsonWebsocketConsumer):
     def notify(self, event):
         channel_name = event["channel_name"]
         instance_pk = event["instance_pk"]
-        model_action = event["action"]
         model_label = event["model"]
 
         viewset_class = self.registry[model_label]
@@ -138,44 +138,15 @@ class SubscriptionConsumer(JsonWebsocketConsumer):
             viewset = viewset_class.from_scope(
                 self.actions[request_id], self.scope, self.kwargs[request_id]
             )
-            broadcast_data = viewset.get_data_to_broadcast(instance_pk, self.pks[request_id])
+            broadcast_data = viewset.get_data_to_broadcast(
+                instance_pk, self.owned_instance_pks[request_id]
+            )
 
             if broadcast_data is not None:
                 instance_data, renderer, action = broadcast_data
-                self.pks[request_id].add(instance_pk)
-                self.send_broadcast(request_id, model_label, action, instance_data, renderer)
-
                 # We don't need to check for membership since it's implicit given broadcast_data isn't None.
                 if action == DELETED:
-                    self.pks[request_id].remove(instance_pk)
-
-
-class RealtimeRouter:
-    def __init__(self, uid="default"):
-        self.registry: Dict[str, Type[RealtimeMixin]] = dict()
-        self.uid = uid
-
-    def register_all(self, viewsets):
-        for viewset in viewsets:
-            self.register(viewset)
-
-    def register(self, viewset):
-        if not hasattr(viewset, "register_realtime"):
-            raise RuntimeError(
-                f"View {viewset.__name__}"
-                "passed to RealtimeRouter does not have RealtimeMixin applied."
-            )
-        label = viewset.register_realtime(self.uid)
-        if label in self.registry:
-            raise RuntimeWarning("You should not register two realitime views for the same model.")
-
-        self.registry[label] = viewset
-
-    def as_consumer(self):
-        # Create a subclass of `SubscriptionConsumer` where the consumer's model
-        # registry is set to this router's registry. Basically a subclass inside a closure.
-        return type(
-            "BoundSubscriptionConsumer",
-            (SubscriptionConsumer,),
-            dict(registry=self.registry),
-        )
+                    self.owned_instance_pks[request_id].remove(instance_pk)
+                else:
+                    self.owned_instance_pks[request_id].add(instance_pk)
+                self.send_broadcast(request_id, model_label, action, instance_data, renderer)
